buat data timbangan baru sudah oke, namun implementasi upload gambar belum

+++ app/src/main/java/com/bangkit/scalesappmobile/domain/model/Scales.kt
@@ -13,7 +13,7 @@ data class Scales(
     @SerializedName("calibrationPeriod")
     val calibrationPeriod: Int,
     @SerializedName("calibrationPeriodInYears")
-    val calibrationPeriodInYears: Int,
+    val calibrationPeriodInYears: Double,
     @SerializedName("equipmentDescription")
     val equipmentDescription: String,
     @SerializedName("id")
