+++ app/src/main/java/com/bangkit/scalesappmobile/domain/model/ScalesDetails.kt
@@ -13,7 +13,7 @@ data class ScalesDetails(
     @SerializedName("calibrationPeriod")
     val calibrationPeriod: Int,
     @SerializedName("calibrationPeriodInYears")
-    val calibrationPeriodInYears: Int,
+    val calibrationPeriodInYears: Double,
     @SerializedName("equipmentDescription")
     val equipmentDescription: String,
     @SerializedName("forms")
